ui-charts: move pickingState to useCanvas() out of BaseChartContextType
The picking state has nothing chart-specific (e.g. tied to space
time chart), it's generic. Thus it doesn't need to be re-implemented
by each individual chart. Move it to the shared useCanvas() hook.
